+++ tests/test_vdecorator.py
@@ -1,7 +1,8 @@
 import pytest
 
-from covador import schema, dpass
-from covador.types import *
+from covador import list_schema, schema, dpass, soft_map, typed_map
+from covador.types import Int, SoftListMap
+from covador.utils import wrap_in
 from covador.vdecorator import (ErrorHandler, ErrorContext,
                                 ValidationDecorator, mergeof)
 
@@ -38,17 +39,17 @@ def test_error_handler():
 
 
 def test_make_validator():
-    schema = lambda: lambda it: {'foo': it + 2}
+    schema = lambda it: {'foo': it + 2}
+    top_schema = lambda: schema
     getter = lambda it: it + 1
-    v = ValidationDecorator(getter, None, schema)
+    v = ValidationDecorator(getter, None, top_schema)
 
     @v()
     def boo(arg, foo):
-        assert arg == 10
-        assert foo == 13
+        return arg, foo
 
-    boo(10)
-    assert boo.schema
+    assert boo(10) == (10, 13)
+    assert boo.schema is schema
 
 
 def test_make_validator_with_skip():
@@ -58,11 +59,9 @@ def test_make_validator_with_skip():
 
     @v()
     def boo(request, arg, foo):
-        assert request == 'request'
-        assert arg == 10
-        assert foo == 13
+        return request, arg, foo
 
-    boo('request', 10)
+    assert boo('request', 10) == ('request', 10, 13)
 
 
 def test_make_validator_with_empty_error_handler():
@@ -102,11 +101,25 @@ def test_validator_with_pipe():
 
     @dpass(params(arg=int) | (lambda r: {'arg': r['arg'] + 10}))
     def boo(request, arg):
-        boo.called = True
-        assert arg == 11
+        return arg
 
-    boo({'arg': 1})
-    assert boo.called
+    assert boo({'arg': 1}) == 11
+
+
+def test_validator_with_piped_schema():
+    getter = lambda it: it
+
+    v = ValidationDecorator(getter, None, schema)
+    @v(soft_map(arg=int) | wrap_in('data'))
+    def boo(request, data):
+        return data
+    assert boo({'arg': 1, 'foo': 'boo'}) == {'arg': 1, 'foo':'boo'}
+
+    v = ValidationDecorator(getter, None, list_schema)
+    @v(soft_map(arg=int) | wrap_in('data'))
+    def boo(request, data):
+        return data
+    assert boo({'arg': [1], 'foo': ['boo']}) == {'arg': 1, 'foo':'boo'}
 
 
 def test_validator_merge():
@@ -119,7 +132,71 @@ def test_validator_merge():
 
     @v(arg=int)
     def boo(_, arg):
-        boo.called = True
-        assert arg in (2, 3)
+        return arg
 
-    boo([{}, {'arg': 3}])
+    assert boo([{}, {'arg': 3}]) == 3
+
+
+def test_validator_soft_map_merge():
+    getter0 = lambda it: it[0]
+    getter1 = lambda it: it[1]
+
+    v0 = ValidationDecorator(getter0, None, schema)
+    v1 = ValidationDecorator(getter1, None, schema)
+    v = mergeof(v0, v1)
+
+    @v(soft_map(), arg=int)
+    def boo(_, arg, boo):
+        return arg, boo
+
+    assert boo([{'boo': 'foo'}, {'arg': 3}]) == (3, 'foo')
+
+
+def test_ensure_item_getter():
+    getter = lambda ctx: ctx
+    v = ValidationDecorator(getter, None, list_schema)
+
+    @v(schema(arg=int))
+    def boo(_, arg):
+        return arg
+
+    assert boo({'arg': ["1"]}) == 1
+
+
+def test_soft_map():
+    getter = lambda ctx: ctx
+    v = ValidationDecorator(getter, None, schema)
+
+    @v(soft_map(arg=int))
+    def boo(_, arg, **data):
+        return arg, data
+
+    assert boo({'arg': '10', 'foo': 'bar'}) == (10, {'foo': 'bar'})
+
+
+def test_soft_list_map():
+    getter = lambda ctx: ctx
+    v = ValidationDecorator(getter, None, list_schema)
+
+    @v(SoftListMap(dict(arg=int)))
+    def boo(_, arg, **data):
+        return arg, data
+
+    assert boo({'arg': ['10'], 'foo': ['bar']}) == (10, {'foo': ['bar']})
+
+
+def test_typed_map():
+    getter = lambda ctx: ctx
+    v = ValidationDecorator(getter, None, schema)
+
+    @v(typed_map(str, int))
+    def boo(_, arg, foo):
+        return arg, foo
+
+    assert boo({'arg': '10', 'foo': 20}) == (10, 20)
+
+    @v(typed_map(str, int) | wrap_in('data'))
+    def boo(_, data):
+        return data
+
+    assert boo({'arg': '10', 'foo': 20}) == {'arg': 10, 'foo': 20}